src/api/requirements.txt
# API Dependencies
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0
azure-identity>=1.15.0
azure-ai-projects>=1.0.0b1
